moved AbstractSimulationWorkflowConfiguration to de.uka.ipd.sdq.codegen.simucontroller.core
