+++ bundles/de.uka.ipd.sdq.codegen.simucontroller.core/META-INF/MANIFEST.MF
@@ -3,6 +3,11 @@ Bundle-ManifestVersion: 2
 Bundle-Name: Simu Controller Core Plug-in
 Bundle-SymbolicName: de.uka.ipd.sdq.codegen.simucontroller.core
 Bundle-Version: 5.2.1.qualifier
+Export-Package: de.uka.ipd.sdq.codegen.simucontroller.core.runconfig
+Require-Bundle: de.uka.ipd.sdq.workflow;bundle-version="5.2.1",
+ de.uka.ipd.sdq.simulation.core;bundle-version="5.2.1",
+ org.palladiosimulator.analyzer.workflow.core;bundle-version="5.2.1",
+ de.uka.ipd.sdq.workflow.launchconfig.core;bundle-version="5.2.1"
 Bundle-ActivationPolicy: lazy
 Bundle-Vendor: sdq.ipd.uka.de
 Bundle-RequiredExecutionEnvironment: JavaSE-17
