+++ bundles/de.uka.ipd.sdq.codegen.simucontroller/src/de/uka/ipd/sdq/pcm/transformations/ApplyConnectorCompletionsJob.java
@@ -27,7 +27,7 @@ import org.palladiosimulator.pcm.resourceenvironment.ResourceenvironmentPackage;
 
 import com.google.common.collect.Streams;
 
-import de.uka.ipd.sdq.codegen.simucontroller.runconfig.AbstractSimulationWorkflowConfiguration;
+import de.uka.ipd.sdq.codegen.simucontroller.core.runconfig.AbstractSimulationWorkflowConfiguration;
 import de.uka.ipd.sdq.featureconfig.Configuration;
 import de.uka.ipd.sdq.pcm.transformations.builder.IBuilder;
 import de.uka.ipd.sdq.pcm.transformations.builder.connectors.ConnectorReplacingBuilder;
